+++ changelog.md
@@ -16,6 +16,7 @@ Codion Change Log
 - DefaultEntityEditModel.DefaultSettings minor refactoring.
 - EntityEditModel.applyEdit() entities Collection parameter replaced with a single entity.
 - DefaultEntityEditor now uses Objects.deepEquals() when notifying.
+- EntityEditModel.applyEdit() replaced with the more focused EntityEditor.EditorValue.propagate() and Propagator.
 ### is.codion.swing.framework.model
 - EntityComboBoxModel.PERSISTENCE_EVENTS renamed PERSISTENCE_AWARE, Builder.persistenceEvents() renamed persistenceAware().
 ### is.codion.swing.framework.ui